project-management-env/pm_env/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
atomicwrites==1.4.0
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==36.0.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
dj-rest-auth==1.1.0
Django==4.0.3
django-allauth==0.42.0
django-extensions==3.1.5
django-guardian==2.4.0
django-rest-knox==4.2.0
djangorestframework==3.11.2
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.4.2
fonttools==4.33.3
idna==3.3
iniconfig==1.1.1
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
pipenv==2022.1.8
platformdirs==2.4.1
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
psycopg2==2.9.3
pure-eval==0.2.2
py==1.11.0
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==7.1.1
pytest-django==4.5.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pywin32==303
pywinpty==2.0.5
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
terminado==0.13.3
tinycss2==1.1.1
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
tzdata==2021.5
urllib3==1.26.9
virtualenv==20.13.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
wget==3.2
widgetsnbextension==3.6.0
